Add validations, switch product stock to int
Introduce input validation and UI limits across multiple forms, and switch product stock types to integers. Key changes:

- Add Regex validations and MaxLength limits for names and search boxes in Beneficiarios, Categorias, Productos, TasaConsumo, Unidades, and Historial to improve input hygiene.
- Adjust numeric controls: limit ranges/decimal places (productos max/stock to integers, dias reposicion limit), update tooltips accordingly.
- Change Producto model: Stock and MaximoPorEntrega from double to int; update ProductoServicio, CSV importer, and related methods to use int for stock and max per delivery.
- Remove Observaciones handling from distribution UI and from ConfirmarDistribucion/Orden insertion (observations column no longer used), and remove the observations textbox from FormDistribucion.
- Minor UI text tweak for product low-stock label.

These changes enforce stricter data integrity, simplify stock handling, and clean up the order/distribution flow.

+++ Modelos/Producto.cs
@@ -6,8 +6,8 @@ namespace ProyectoCatedra.Modelos
         public string SKU { get; set; } = string.Empty;
         public string Nombre { get; set; } = string.Empty;
         public int IdCategoria { get; set; }
-        public double Stock { get; set; }
-        public double? MaximoPorEntrega { get; set; }
+        public int Stock { get; set; }
+        public int? MaximoPorEntrega { get; set; }
         public int? DiasReposicion { get; set; }
         
         // Propiedad extra para mostrar en la UI sin hacer joins complejos
